+++ pymilo/transporters/tree_transporter.py
@@ -68,7 +68,12 @@ class TreeTransporter(AbstractTransporter):
         """
         content = data[key]
 
-        if (key == "tree_" and (model_type == "DecisionTreeRegressor")):
+        if (key == "tree_" and 
+            (model_type == "DecisionTreeRegressor" 
+             or model_type == "DecisionTreeClassifier"
+             or model_type == "ExtraTreeRegressor" 
+             or model_type == "ExtraTreeClassifier"
+             )):
             gdst = GeneralDataStructureTransporter()
             tree_params = content['params']